+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [0.6.2] - 2021-05-26
+
 ### Changed
 
 - Fix `iframe-has-title` such that if the attribute is ommitted completely that it fires correctly.
@@ -91,7 +93,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - Initial release 🎉.
 
-[unreleased]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.6.0...HEAD
+[unreleased]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.6.2...HEAD
+[0.6.2]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.6.1...v0.6.2
+[0.6.1]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/vue-a11y/eslint-plugin-vuejs-accessibility/compare/v0.4.0...v0.5.0